+++ Source/VerveAdapterRewarded.swift
@@ -1,4 +1,4 @@
-// Copyright 2023-2023 Chartboost, Inc.
+// Copyright 2023-2024 Chartboost, Inc.
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file.
@@ -31,7 +31,7 @@ final class VerveAdapterRewardedAd: VerveAdapterAd, PartnerAd {
         ad = hyBidAd
         // Load differently depending on whether this is a bidding or non-programatic ad
         if let adm = request.adm {
-            hyBidAd.prepareCustomMarkupFrom(adm)
+            hyBidAd.prepareAdWithContent(adContent: adm)
         } else {
             hyBidAd.load()
         }