+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Note the first digit of every adapter version corresponds to the major version of the Chartboost Mediation SDK compatible with that adapter. 
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
+### 4.2.20.0.0
+- This version of the adapter has been certified with HyBid 2.20.0.
+
 ### 4.2.18.1.1
 - Add support for Adaptive Banners.
 - This version of the adapters has been certified with HyBid 2.18.1.